+++ base/include/luminous_intensity.hpp
@@ -7,8 +7,8 @@ namespace sakurajin{
         namespace base{
             class luminous_intensity : public unit_t<7>{
             public:
-                luminous_intensity(double val);
-                luminous_intensity(double val, double mult);
+                luminous_intensity(long double val);
+                luminous_intensity(long double val, long double mult);
                 
             };
