+++ base/include/temperature.hpp
@@ -7,8 +7,8 @@ namespace sakurajin{
         namespace base{
             class temperature : public unit_t<4>{
             public:
-                temperature(double val);
-                temperature(double val, double mult);
+                temperature(long double val);
+                temperature(long double val, long double mult);
                 
             };
