+++ base/src/temperature.cpp
@@ -2,5 +2,5 @@
 
 using namespace sakurajin::unit_system::base;
 
-temperature::temperature(double val): unit_t<4>(val){}
-temperature::temperature(double val, double mult): unit_t<4>(val,mult){}
+temperature::temperature(long double val): unit_t<4>(val){}
+temperature::temperature(long double val, long double mult): unit_t<4>(val,mult){}
